geoscience/speed-and-structure/Team_AYOUA/requirements.txt
anytree==2.13.0
fvcore==0.1.5.post20221221
hydra-core==1.3.2
# numpy==2.3.2
omegaconf==2.3.0
pandas==2.3.1
pytorch_lightning==2.5.0.post0
scikit_learn==1.7.1
timm==1.0.19
torch==2.5.1
torch_ema==0.3
torch_optimizer==0.3.0
torchmetrics==1.6.1
torchvision==0.20.1
transformers==4.48.0
